// src/app/(app)/recommendations/components/type.ts
export type Bazin = {
	ativo: string
	cotacao: number
	evebit: number
	roic: number
	liqtwomeses: number
	setor: string
	subsetor: string
	segmento: string
	rankevebit: number
	rankroic: number
	magicformula: number
	is_prov: string
	dy: string
	divporano: string
	precojusto: string
	precoatual: string
	sobresubvalorizacao: string
	situacao: 'SUBVALORIZADO' | 'SOBREVALORIZADO'
	bazin: string
	rankingbazin: number
}